Add functionality that adds value to dynamic string (#42)
This means e.g. overlay widget gauges (in a camera with support for
that) can display the digital value on screen. Along with this update,
we also move to the current version of native ACAP 4 SDK for building
and hence drop support for legacy versions of Axis OS.

For legacy devices, the previous commit can still be successfully used,
only that it (at this point) will not have the data written to the
dynamic string.

Signed-off-by: Joakim Roubert <joakimr@axis.com>